WIP - fork dex server in to app
This is a WIP on taking dex's actual OIDC server code, and folding it in to our
app.

I'm parking this for a moment though, because I'm wondering if the benefits of
having a "library" quality OIDC server outweigh just vendoring dex, and hacking
around it.  The latter might make it easier to get updates to the code over
time.