Add browser-based smoke test
Testing the full webauthn flow is hard programmatically because it involves
javascript and authenticators, and it tedious manually. So use Chrome
automation to add a smoke test that exercices the add user/enroll/login flow.

To support this, the enroll user function was refactored a little to pass
output up, which main displays

This can be run with `TEST_E2E=1 go test -run=TestE2E -v -count=1 .`

Add it to CI as well, to make sure things generally work.