Re-work CLI, drop migration + backwards compat
Move to kong for the CLI. Still not the biggest fan, but it allows for a lot of
boilerplate to be removed. Once the admin commands are removed, might consider
dropping it.

Remove the backwards compatability/migration code. This cleans up a bunch of
stuff that was annoying to maintain, and was getting less tested as time goes
on. Users can deploy the last version prior to this, which will migrate data in
a way that is forwards compatible. This can then be deployed on top, with an
updated config.

Drop multi-tenancy for the minute. This might have worked, but was not actually
tested. Nothing stops us from bringing it back and the code will be developed
to support this, but not having it right now will ease the migration from CLI
admin commands to a real web UI.