Support dynamic client registration
This adds support for dynamic OIDC client registration. This lets us meet the
goal of having services obtain an identity for a user, without requiring a
pre-registered client. It registers the clients in the DB, with an expiration
deadline.

This is a very basic implementation, and has plenty of rough edges to clean up.
It's good enough for the current use, but we really should add scope checking
and consent before we hand out info to a random client, and probably pairwise
subjects.